requirements.txt
asgiref==3.3.4
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.1
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-crispy-forms==1.11.2
django-qr-creator==0.1.5
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.3.0
